requirements.txt
django-cors-headers==3.4.0
djangorestframework==3.13.1
pytz==2020.1
sqlparse==0.4.2
